Module for computing incidence rates for cohorts against an OMOP CDM

Computes incidence rates for cohorts against the OMOP CDM

Super class

Strategus::StrategusModule -> CohortIncidenceModule

Public fields

tablePrefix

The table prefix to append to results tables

Methods

Public methods

CohortIncidenceModule$new()

CohortIncidenceModule$execute()

CohortIncidenceModule$createResultsDataModel()

CohortIncidenceModule$getResultsDataModelSpecification()

CohortIncidenceModule$uploadResults()

CohortIncidenceModule$createModuleSpecifications()

CohortIncidenceModule$validateModuleSpecifications()

CohortIncidenceModule$clone()

Inherited methods

Strategus::StrategusModule$createSharedResourcesSpecifications()

Strategus::StrategusModule$validateSharedResourcesSpecifications()

Method new()

Initialize the module

Usage

CohortIncidenceModule$new()

Method execute()

Execute the CohortIncidence package

Usage

CohortIncidenceModule$execute(
  connectionDetails,
  analysisSpecifications,
  executionSettings
)

Arguments

connectionDetails

An object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails() function.

analysisSpecifications

An object of type AnalysisSpecifications as created by createEmptyAnalysisSpecificiations().

analysisSpecifications

An object of type AnalysisSpecifications as created by createEmptyAnalysisSpecificiations().

executionSettings

An object of type ExecutionSettings as created by createCdmExecutionSettings() or createResultsExecutionSettings().

Method createResultsDataModel()

Create the results data model for the module

Usage

CohortIncidenceModule$createResultsDataModel(
  resultsConnectionDetails,
  resultsDatabaseSchema,
  tablePrefix = ""
)

Arguments

resultsConnectionDetails

The connection details to the results database which is an object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails() function.

resultsConnectionDetails

The connection details to the results database which is an object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails() function.

resultsDatabaseSchema

The schema in the results database that holds the results data model.

tablePrefix

A prefix to apply to the database table names (optional).

tablePrefix

A prefix to apply to the database table names (optional).

Method getResultsDataModelSpecification()

Get the results data model specification for the module

Usage

CohortIncidenceModule$getResultsDataModelSpecification(tablePrefix = "")

Arguments

tablePrefix

A prefix to apply to the database table names (optional).

tablePrefix

A prefix to apply to the database table names (optional).

Method uploadResults()

Upload the results for the module

Usage

CohortIncidenceModule$uploadResults(
  resultsConnectionDetails,
  analysisSpecifications,
  resultsDataModelSettings
)

Arguments

resultsConnectionDetails

The connection details to the results database which is an object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails() function.

resultsConnectionDetails

The connection details to the results database which is an object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails() function.

analysisSpecifications

An object of type AnalysisSpecifications as created by createEmptyAnalysisSpecificiations().

analysisSpecifications

An object of type AnalysisSpecifications as created by createEmptyAnalysisSpecificiations().

resultsDataModelSettings

The results data model settings as created using [@seealso createResultsDataModelSettings()]

Method createModuleSpecifications()

Creates the CohortIncidence Module Specifications

Usage

CohortIncidenceModule$createModuleSpecifications(irDesign = NULL)

Arguments

irDesign

The incidence rate design created from the CohortIncidence package

Method validateModuleSpecifications()

Validate the module specifications

Usage

CohortIncidenceModule$validateModuleSpecifications(moduleSpecifications)

Arguments

moduleSpecifications

The CohortIncidence module specifications

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortIncidenceModule$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.