createResultDataModels {Strategus}	R Documentation

Create Result Data Models

Description

Use this at the study design stage to create data models for modules This functions loads modules and executes any custom code to create schemas in a results database If recreate is set to TRUE all existing data will be removed, otherwise

Usage

createResultDataModels(
  analysisSpecifications,
  executionSettings,
  executionScriptFolder = NULL,
  keyringName = NULL
)

Arguments

analysisSpecifications
An object of type AnalysisSpecifications as created by createEmptyAnalysisSpecificiations().

executionSettings
An object of type ExecutionSettings as created by createCdmExecutionSettings() or createResultsExecutionSettings().

executionScriptFolder
Optional: the path to use for storing the execution script. when NULL, this function will use a temporary file location to create the script to execute.

keyringName
The name of the keyring to operate on. This function assumes you have created the keyring before calling this function. It defaults to NULL to select the default keyring. If the keyring is password protected, the password must be stored in the environment variable STRATEGUS_KEYRING_PASSWORD so it is retrieved using the command Sys.getenv("STRATEGUS_KEYRING_PASSWORD")

[Package Strategus version 0.0.6 Index]