ensureAllModulesInstantiated {Strategus}	R Documentation

Ensure all modules are instantiated

Description

Ensure that all modules referenced in the analysis specifications are instantiated locally in the folder specified in the INSTANTIATED_MODULES_FOLDER environmental variable.

Missing modules will be fetched from remote repositories.

This function will also check whether there are different versions of the same module specified, which is not allowed, and whether all modules required by the specified modules are also instantiated.

Usage

ensureAllModulesInstantiated(analysisSpecifications)

Arguments

analysisSpecifications
An object of type AnalysisSpecifications as created by createEmptyAnalysisSpecificiations().

Value

A tibble listing the instantiated modules.

[Package Strategus version 0.0.6 Index]