getModuleList {Strategus}	R Documentation

Provides a list of HADES modules to run through Strategus

Description

This function provides a list of modules and their locations that may be used with Strategus.

Usage

getModuleList()

Value

A data.frame() of modules that work with Strategus. This will contain: module = The name of the module version = The version of the module remote_repo = The remote location of the module (i.e. github.com) remote_username = The organization of the module (i.e. OHDSI) module_type = 'cdm' or 'results'. 'cdm' refers to modules that are designed to work against patient level data in the OMOP CDM format. 'results' refers to modules that are designed to work against a results database containing output from a 'cdm' module.

[Package Strategus version 0.0.6 Index]