runSchemaCreation {Strategus}	R Documentation

Results upload callbacks for inserting results in to a database

Description

Results upload callbacks for inserting results in to a database

Usage

runSchemaCreation(
  analysisSpecifications,
  keyringSettings,
  moduleIndex,
  executionSettings,
  ...
)

[Package Strategus version 0.0.6 Index]