storeConnectionDetails {Strategus}	R Documentation

Store connection details in a secure location

Description

Store connection details in a secure location

Usage

storeConnectionDetails(
  connectionDetails,
  connectionDetailsReference,
  keyringName = NULL
)

Arguments

connectionDetails
An object of type connectionDetails as created by the DatabaseConnector::createConnectionDetails() function.

connectionDetailsReference
A string that can be used to retrieve the settings from the secure store.

keyringName
The name of the keyring to operate on. This function assumes you have created the keyring before calling this function. It defaults to NULL to select the default keyring. If the keyring is password protected, the password must be stored in the environment variable STRATEGUS_KEYRING_PASSWORD so it is retrieved using the command Sys.getenv("STRATEGUS_KEYRING_PASSWORD")

Value

Does not return anything. Is called for the side effect of having the connection details stored.

See Also

retrieveConnectionDetails()

[Package Strategus version 0.0.6 Index]