Load module execution space inside and renv inspired by targets::tar_script but allowing custom variable execution

Designed to allow more human readable code that is executed inside a module as well as simple variable substituion for injecting constants (e.g. simple parameters or file paths used inside and outside of modules)

withModuleRenv(
  code,
  moduleFolder,
  injectVars = list(),
  tempScriptFile = tempfile(fileext = ".R"),
  useLocalStrategusLibrary = TRUE,
  job = FALSE,
  processName = paste(moduleFolder, "_renv_run")
)

Arguments

code

code block to execute

moduleFolder

Instantiated Strategus module folder

injectVars

list of var names list(name=value) to replace (e.g. replace list(foo = "some string") will find the pattern foo and replace it with the string some string - be careful!

tempScriptFile

tempFile to write script to

useLocalStrategusLibrary

Use the locally installed Strategus library? TRUE will use the Strategus installation from the calling R process.

job

run as rstudio job

processName

String name for process

Value

NULL invisibly

Details

This pattern also allows dependency injection which could be used if you don't want to use and renv and (instead) would like to use docker images or just execution in the base environment for testing/debugging