createCdmExecutionSettings {Strategus}	R Documentation

Create CDM execution settings

Description

Create CDM execution settings

Usage

createCdmExecutionSettings(
  connectionDetailsReference,
  workDatabaseSchema,
  cdmDatabaseSchema,
  cohortTableNames = CohortGenerator::getCohortTableNames(cohortTable = "cohort"),
  workFolder,
  resultsFolder,
  minCellCount = 5,
  resultsConnectionDetailsReference = NULL,
  resultsDatabaseSchema = NULL
)

Arguments

connectionDetailsReference
A string that can be used to retrieve database connection details from a secure local store.

workDatabaseSchema
A database schema where intermediate data can be stored. The user (as identified in the connection details) will need to have write access to this database schema.

cdmDatabaseSchema
The database schema containing the data in CDM format. The user (as identified in the connection details) will need to have read access to this database schema.

cohortTableNames
An object identifying the various cohort table names that will be created in the workDatabaseSchema. This object can be created using the CohortGenerator::getCohortTableNames() function.

workFolder
A folder in the local file system where intermediate results can be written.

resultsFolder
A folder in the local file system where the module output will be written.

minCellCount
The minimum number of subjects contributing to a count before it can be included in results.

resultsConnectionDetailsReference
A string that can be used to retrieve the results database connection details from a secure local store.

resultsDatabaseSchema
A schema where the results tables are stored

Value

An object of type ExecutionSettings.

[Package Strategus version 0.0.6 Index]